requirements.txt
# VokaFlow Backend Dependencies
# Production-ready FastAPI backend with AI capabilities

# ===== CORE FRAMEWORK =====
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# ===== ASYNC & HTTP =====
httpx==0.25.2
aiofiles==23.2.1
aioredis==2.0.1
asyncio-mqtt==0.16.1

# ===== DATABASE =====
sqlalchemy==2.0.23
databases[postgresql,sqlite]==0.8.0
psycopg2-binary==2.9.9
alembic==1.13.1

# ===== REDIS =====
redis==5.0.1
hiredis==2.2.3

# ===== AUTHENTICATION & SECURITY =====
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
cryptography==41.0.7

# ===== AI & ML =====
openai==1.3.5
transformers==4.36.0
torch==2.1.1
numpy==1.24.4
scipy==1.11.4

# ===== TRANSLATION =====
googletrans==4.0.0rc1
deep-translator==1.11.4
langdetect==1.0.9

# ===== SPEECH PROCESSING =====
speechrecognition==3.10.0
pydub==0.25.1
librosa==0.10.1
soundfile==0.12.1

# ===== COMPUTER VISION =====
opencv-python==4.8.1.78
opencv-contrib-python==4.8.1.78
mediapipe==0.10.8
Pillow==10.1.0

# ===== KINECT =====
# freenect2==0.2.0  # Uncomment if using Kinect

# ===== WEB SCRAPING & HTTP =====
requests==2.31.0
beautifulsoup4==4.12.2
selenium==4.15.2

# ===== TASK QUEUE & BACKGROUND JOBS =====
celery==5.3.4
kombu==5.3.4

# ===== MONITORING & LOGGING =====
prometheus-client==0.19.0
structlog==23.2.0
loguru==0.7.2

# ===== DATA PROCESSING =====
pandas==2.1.3
pyarrow==14.0.1
polars==0.19.14

# ===== CONFIGURATION =====
python-dotenv==1.0.0
pyyaml==6.0.1
toml==0.10.2

# ===== UTILITIES =====
python-dateutil==2.8.2
pytz==2023.3
click==8.1.7
rich==13.7.0
typer==0.9.0

# ===== TESTING =====
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2  # For testing FastAPI

# ===== CODE QUALITY =====
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1

# ===== DOCUMENTATION =====
mkdocs==1.5.3
mkdocs-material==9.4.8

# ===== DEVELOPMENT =====
pre-commit==3.5.0
ipython==8.17.2
jupyter==1.0.0

# ===== FILE HANDLING =====
openpyxl==3.1.2
python-magic==0.4.27
filetype==1.2.0

# ===== EMAIL & NOTIFICATIONS =====
smtplib==3.11.0  # Built-in
emails==0.6.0

# ===== WEBSOCKETS =====
websockets==12.0

# ===== CACHING =====
cachetools==5.3.2

# ===== SYSTEM MONITORING =====
psutil==5.9.6
py-cpuinfo==9.0.0

# ===== TIMEZONE =====
pytz==2023.3

# ===== JSON & DATA =====
orjson==3.9.10
ujson==5.8.0

# ===== ENVIRONMENT =====
python-environ==0.4.54

# ===== CORS =====
# Included in FastAPI

# ===== PRODUCTION DEPENDENCIES =====
gunicorn==21.2.0  # Alternative to Uvicorn for production
supervisor==4.2.5  # Process management

# ===== OPTIONAL: GPU ACCELERATION =====
# torch==2.1.1+cu118 --index-url https://download.pytorch.org/whl/cu118
# torchvision==0.16.1+cu118 --index-url https://download.pytorch.org/whl/cu118
# torchaudio==2.1.1+cu118 --index-url https://download.pytorch.org/whl/cu118
